Move load_data to utils
